requirements.txt
# ── LLM providers (install whichever matches your LLM_PROVIDER in .env) ──────
openai>=1.0.0                  # LLM_PROVIDER=openai  (ChatGPT / GPT-4o)
anthropic>=0.25.0              # LLM_PROVIDER=anthropic (Claude)
google-generativeai>=0.7.0     # LLM_PROVIDER=google  (Gemini)

# ── Google APIs ───────────────────────────────────────────────────────────────
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.126.0

# ── Web search ────────────────────────────────────────────────────────────────
tavily-python>=0.3.3

# ── Scheduler ─────────────────────────────────────────────────────────────────
APScheduler>=3.10.4

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.1.0
pytest-mock>=3.14.0
